Update Note #7
- ProductDetail Entity added
- ProductDetailRepository interface added
- ProductDetailRepositoryTest added
- If a bidirectional relationship is established through @OneToOne(mappedBy = ""), a StackOverflowError occurs because a circular reference occurs when using ToString. Exclude from ToString using exclude to remove circular references.
ex)
@OneToOne(mappedBy = "")
@ToString.Exclude

- Provider Entity & ProviderRepository creation and testing
@OneToMany(mappedBy = "", fetch = FetchType.EAGER)
*fetchType.EAGER: @OneToMany's default fetch strategy is Lazy, so it is adjusted to immediate loading. If the lazy loading method is used in the test, an error occurs as 'no Session'.
@JoinColumn(name = "provider_id")
@ToString.Exclude

- Category Entity & CategoryRepository creation and testing
In Category Entity,
@OneToMany(fetch = FetchType.EAGER)
@JoinColumn(name = "")
If you use the above two, a one-to-many unidirectional relationship is mapped without any separate settings in the other entity. @JoinColumn is not required, but if it is not present, a strategy of creating a join table as an intermediate table is adopted.
The disadvantage of the OneToMany unidirectional relationship is that a foreign key is added to the partner table, not the subject of the mapping. Unlike the ManyToOne structure, this method generates an update query for the relative table to set a foreign key.

- Add Producer entity, prepare many-to-many one-way mapping
- Add and test ProducerRepositoryTest
- Add and test cascade
- Added and tested orphanRemoval